Add files via upload
Jupyter notebook file with workings on WP2. Regex and webscraping was carried out by Oran. I was tasked with making a functional dataframe/CSV file from a list of over 1000 medicines. This involved manually removing medicines that do not fall under the category of small drug molecules (defined as MW of less than 1000 (the definition taken in Delaney et al. (2004)), the creating of a function to remove miscellaneous families and general cleaning of the data before handing over to Isobel.