Add files via upload
This is a recreation of the code I ran in November when Oran and I were able to diagnose what had gone wrong and why following the GitHub DeepChem tutorials did not work. Unfortunately, due to the lack of cross validation these results cannot be compared with the results created later on in the project. If time had allowed I could have tried to apply a function defined by Oran to this to cross validate and allow statistical comparison. The important thing to note is that the epochs could be changed and some hyperparameter optimisation could be undertaken by passing a list rather than a single value. However, size of neural network (number of dense layers and number of nodes within a dense layer) are not flexible with this approach. A low level approach, following the same set of GitHub DeepChem tutorials was attempted but did not yield any success.