GCNN update
The code in this jupyter notebook recreates work undertaken in November that was overwritten due to OneDrive syncing issues and follows deepchem github tutorials. Once Oran and I had solved why deepchem was not functioning I set out to try and get predictive scores out of a deepchem model. It is worth noting that some parameters such as epoch could be optimised whilst others such as dense layer number / size could not. As we realised this system did not allow attenuation of datasets due to errors in the deepchem CSV loader function and batch normalization, further investigation of hyperparameter optimization was sidelined.